+++ wiki/Changelog.md
@@ -10,9 +10,13 @@ Terminal features
 Misc
   * Support fallback to legacy wslbridge for old Windows versions (mintty/wsltty#350).
 
+Desktop integration
+  * Avoid taskbar icon width twitching in Windows 11 (#1263).
+
 Configuration
   * New option AllowPasteSelection (#1264).
   * Drop AllowSetSelection from interactive Options menu (~#1264).
+  * New option TabFont.
 
 ### 3.7.1 (24 February 2024) ###